Refactor: move API to comments app
